move: Move app to src
